+++ src/SparseDiffTools.jl
@@ -5,7 +5,6 @@ using FiniteDiff
 using ForwardDiff
 using Graphs
 using Graphs: SimpleGraph
-using Requires
 using VertexSafeGraphs
 using Adapt
 
@@ -69,16 +68,16 @@ Base.@pure __parameterless_type(T) = Base.typename(T).wrapper
 parameterless_type(x) = parameterless_type(typeof(x))
 parameterless_type(x::Type) = __parameterless_type(x)
 
-function __init__()
-    @require Zygote="e88e6eb3-aa80-5325-afca-941959d7151f" begin
-        export numback_hesvec, numback_hesvec!,
-               autoback_hesvec, autoback_hesvec!,
-               auto_vecjac, auto_vecjac!,
-               ZygoteVecJac, ZygoteHesVec
+#if !isdefined(Base, :get_extension)
+    using Requires
+#end
 
-        include("differentiation/vecjac_products_zygote.jl")
-        include("differentiation/jaches_products_zygote.jl")
-    end
+function __init__()
+    #@static if !isdefined(Base, :get_extension)
+        @require Zygote="e88e6eb3-aa80-5325-afca-941959d7151f" begin
+            include("../ext/SparseDiffToolsZygote.jl")
+        end
+    #end
 end
 
 end # module
