--- src/differentiation/vecjac_products_zygote.jl
@@ -1,14 +1,0 @@
-function auto_vecjac!(du, f, x, v, cache1 = nothing, cache2 = nothing)
-    !hasmethod(f, (typeof(x),)) && error("For inplace function use autodiff = false")
-    du .= reshape(auto_vecjac(f, x, v), size(du))
-end
-
-function auto_vecjac(f, x, v)
-    vv, back = Zygote.pullback(f, x)
-    return vec(back(reshape(v, size(vv)))[1])
-end
-
-#ZygoteVecJac = VecJac
-ZygoteVecJac(args...; autodiff = true, kwargs...) = VecJac(args...; autodiff = autodiff, kwargs...)
-
-#
